log 11/24/18
May have completely misunderstood use of start and finish functions. I copy and pasted their contents within Constructor and Destructor within the files. When I got to one file one of them was trying to reference the initialize function of another file. Unfortunately I realized that the Constructor could not be referenced in this way and couldn't be a substitute for the initialize method. Will probably have to get further input about how these functions operate and may need to start again from the beginning.