Fixed flee behavior
Successfully implemented Flee behavior but came at the cost of the Seek behavior. If Seek is to be implemented Flee needs to be commented out in the XML code and Seek uncommented. More AI behaviors will be added but this will probably be the basis of implementing AI behaviors for our game engine until another efficient method is realized.